+++ textattack/search_methods/__init__.py
@@ -15,4 +15,5 @@ from .genetic_algorithm import GeneticAlgorithm
 from .alzantot_genetic_algorithm import AlzantotGeneticAlgorithm
 from .improved_genetic_algorithm import ImprovedGeneticAlgorithm
 from .particle_swarm_optimization import ParticleSwarmOptimization
+from .particle_swarm_optimization_leap import ParticleSwarmOptimizationLEAP
 from .differential_evolution import DifferentialEvolution